build: make char to be unsigned char on gcc
OS2EMX_PLAIN_CHAR make BYTE to be char, which is signed char by default.
This causes KAISPEC.bSilence to have wrong value when playing 8 bits
audio. On 8 bits audio, bSilence has 0x80, which is -128 in signed char.
So when playing 8 bits audio with soft volume mode, sound is distorted.

To avoid this, make char to be unsigned char.

    modified:   Makefile