uniaud: match ulBufSize to a buffer size passed to a call back
Currently, ulBufSize is not the same as a buffer size passed to a call
back. That is, ulBufSize is pcm->period_size and size for a call back
is pcm->period_size * 2.

In case of DART, both are same. Mimic this. And this makes pre-allocation
of a buffer in according to ulBufSize for a call back convenient.

In practice, this makes pre-allocation of a mixer stream buffer simple
and prevents memory waste because of multiplying ulBufferSize by
ulNumBuffers for UNIAUD mode.

    modified:   kai_uniaud.c