kai: allow to set the minimum samples with KAI_MINSAMPLES
The minimum buffer size is different according to audio cards.

For example, 512 samples is too small for Intel HD Audio. And choppy
sounds are generated. Whereas 512 samples is enough for ICH AC97.

More precisely, a buffer size matters.

Because of this, set the default minimum samples to 2048. And allow to
modify with KAI_MINSAMPLES env. var.

    modified:   kai.c
    modified:   kai.h