kai: check if a stream is playing in kaiStop()
a stream other than a playing stream can stop playback at all.

For exaple, consider two opened streams. One stream is playing, and
the other stream is not playing.

BTW, if a stream not being played is passed to kaiStop(), playback of a
mixer stops completely. Because number of current playing streams is 1.

Check if a streams is playing or not.

    modified:   kai.c