kaidemo: fix very first samples are repeated when replaying on DART mode
If playing is completed, m_ulStatus becomes KAIS_COMPLETED.

When replaying, DART requires at least two buffers. On filling in first
buffer, mmio is sought to the begining because m_ulStatus is
KAIS_COMPLETED. And on filling in seconds buffer, m_ulStatus is still
KAIS_COMPLETED. Therefore, second buffer is filled with the same first
samples.

As a result, first samples are repeated twice.

    modified:   kaidemo.c