dart: fix not playing again without workaround for UNIAUD audio driver
When issuing MCI_STOP in dartStop(), DART sub-system tries to play
remaining buffers. And whenever a buffer is played, MixHandler() is
called. In order, pmixWrite() is called in MixHandler(). After all,
another buffer is filled, again. That is, any buffers are not flushed.

Because of this, when trying to play later, MixHandler() is not called
any more because any new buffers are not filled by pmixWrite() in
dartPlay(). Audio buffers are full already.

Preventing MixHandler() from calling pmixWrite(), fixed this problem
because remaining buffers are really flushed and pmixWrite() in dartPlay()
can kick off playing.

And looping sound or dead lock does not occur any more without this
patch. This problem seemed to be fixed by UNIAUD audio driver itself.

Now, make a normal code path without workaround default.

    modified:   kai_dart.c