kai: allow to override sampling rate of a default soft mixer
If a sound card supports up to 44100 like SB16, libkai would not work
unless KAI_NOSOFTMIXER=1 because sampling rate of a default soft mixer
is 48000.

Using kaiEnableSoftMixer() makes a sampling rate change, but this is
for a programmer not a user.

And use only one variable, pszEnv, for a value of env. vars.

    modified:   kai.c