server: fix dead-lock when calling kaiClose() in MCI_CLOSE of ksoftseq
If MCI_CLOSE of MCD such as ksoftseq, calls MCI_CLOSE in the other
process such as kaisrv, MCI_CLOSE in the other process seems to be
blocked until MCI_CLOSE of MCD is returned.

Therefore, trying to get rc from kaisrv in MCI_CLOSE of ksoftseq leads
to dead-lock.

To avoid this, assume rc will be 0. However, this is enough reasonable
because kaiClose()/kaiMixerClose()/kaiMixerStreamClose() almost succeed
if an instance is valid.

Fixed #11.

    modified:   kai_server.c